chore: move locale-receiver to a standalone dir
