+++ components/action-sheet/index.tsx
@@ -1,4 +1,4 @@
 import ActionSheet from './ActionSheet';
-import LocaleReceiver from '../locale-provider/LocaleReceiver';
+import LocaleReceiver from '../locale-receiver';
 
 export default LocaleReceiver('ActionSheet')(ActionSheet);
