+++ components/alert/index.tsx
@@ -1,4 +1,4 @@
 import Alert from './Alert';
-import LocaleReceiver from '../locale-provider/LocaleReceiver';
+import LocaleReceiver from '../locale-receiver';
 
 export default LocaleReceiver('Alert')(Alert);
