+++ components/confirm/index.tsx
@@ -1,4 +1,4 @@
 import Confirm from './Confirm';
-import LocaleReceiver from '../locale-provider/LocaleReceiver';
+import LocaleReceiver from '../locale-receiver';
 
 export default LocaleReceiver('Confirm')(Confirm);
