+++ components/date-picker-view/index.tsx
@@ -1,4 +1,4 @@
 import DatePickerView from './DatePickerView';
-import LocaleReceiver from '../locale-provider/LocaleReceiver';
+import LocaleReceiver from '../locale-receiver';
 
 export default LocaleReceiver('DatePickerView')(DatePickerView);
