+++ components/date-picker/index.tsx
@@ -1,4 +1,4 @@
 import DatePicker from './DatePicker';
-import LocaleReceiver from '../locale-provider/LocaleReceiver';
+import LocaleReceiver from '../locale-receiver';
 
 export default LocaleReceiver('DatePicker')(DatePicker);
