+++ components/index.tsx
@@ -44,4 +44,5 @@ export { default as Wheel } from './wheel';
 export { default as NavBar } from './nav-bar';
 export { default as Marquee } from './marquee';
 export { default as TabBar } from './tab-bar';
-export { default as LocaleProvider, LocaleReceiver } from './locale-provider';
+export { default as LocaleProvider } from './locale-provider';
+export { default as LocaleReceiver } from './locale-receiver';
