+++ components/keyboard/index.tsx
@@ -1,4 +1,4 @@
 import Keyboard from './Keyboard';
-import LocaleReceiver from '../locale-provider/LocaleReceiver';
+import LocaleReceiver from '../locale-receiver';
 
 export default LocaleReceiver('Keyboard')(Keyboard);
