+++ components/locale-provider/index.tsx
@@ -1,4 +1,3 @@
 import LocaleProvider from './LocaleProvider';
 
-export { default as LocaleReceiver } from './LocaleReceiver';
 export default LocaleProvider;
