+++ components/locale-receiver/index.tsx
@@ -1,0 +1,3 @@
+import LocaleReceiver from './LocaleReceiver';
+
+export default LocaleReceiver;
