+++ components/locale-receiver/style/index.tsx
@@ -1,0 +1,1 @@
+import '../../style';
