+++ components/picker/index.tsx
@@ -1,5 +1,5 @@
 import Picker from './Picker';
-import LocaleReceiver from '../locale-provider/LocaleReceiver';
+import LocaleReceiver from '../locale-receiver';
 // import Stack from './StackPicker';
 
 // Picker.Stack = Stack;
