+++ components/select/index.tsx
@@ -1,4 +1,4 @@
 import Select from './Select';
-import LocaleReceiver from '../locale-provider/LocaleReceiver';
+import LocaleReceiver from '../locale-receiver';
 
 export default LocaleReceiver('Select')(Select);
